Paging on threads.
This changes threads from being a cursor to being an array. This might turn
out to be inefficient. If so, we'll have to change the paging code to
take in a cursor or an array, and if a cursor has been given, just pull in
enough results (using skip and limit) to see how many results there are in
either direction.