TreeForm function.
This isn't exactly correct yet; still trying to find the right structure.

The use case is the one in test_marshal -- I have an array of projects, and 
I want to turn that into a mapping based on some value of each project. I want
the value for each project to be an array of objects, which is mapped from the
areas field of each project.