Don't delete drafts on fetch.
In case a user's browser session crashes, we'd like the draft to stick
around. In addition, in case the draft is live, we have to fetch the slug
from the draft. Instead, we should delete on save.

This cascades into autosave, which created a new draft of whatever the
user was working on. We don't want to create a new draft if we're editing
a draft, since now that would create a *second* draft; instead we just want
to update the draft we're working on. Of course, if we're editing a post,
we still want to create a draft.

I'm not sure this covers all the possible cases; there are certainly a few
bugs in this code.